shared tests between memory and persistent implementations
